+++ test/functional/specs/security-tokens-api.specs.js
@@ -14,10 +14,10 @@ test.describe('security tokens api', function () {
     password: 'foo'
   }
 
-  const moduleUser = {
-    name: 'foo-module-user',
+  const serviceUser = {
+    name: 'foo-service-user',
     role: 'module',
-    email: 'module@foo.com',
+    email: 'service@foo.com',
     password: 'foo'
   }
 
@@ -168,22 +168,22 @@ test.describe('security tokens api', function () {
     }
 
     testRole(operatorUser)
-    testRole(moduleUser)
+    testRole(serviceUser)
     testRole(pluginUser)
     testRole(serviceRegistererUser)
 
     test.describe('When user has service-registerer role and request tokens of an user different to himself', () => {
       let operatorUserId
-      let moduleUserId
+      let serviceUserId
 
       test.before(() => {
         return utils.ensureUser(authenticator, operatorUser)
           .then(() => {
-            utils.ensureUser(authenticator, moduleUser)
+            utils.ensureUser(authenticator, serviceUser)
           }).then(() => {
             return getUsers().then(usersResponse => {
               operatorUserId = usersResponse.body.find(userData => userData.name === operatorUser.name)._id
-              moduleUserId = usersResponse.body.find(userData => userData.name === moduleUser.name)._id
+              serviceUserId = usersResponse.body.find(userData => userData.name === serviceUser.name)._id
               return utils.ensureUserAndDoLogin(authenticator, serviceRegistererUser)
             })
           })
@@ -204,7 +204,7 @@ test.describe('security tokens api', function () {
       test.it('should return tokens data if requested user has "module" role and requested type is "apiKey"', () => {
         return getAuthTokens({
           type: 'apiKey',
-          user: moduleUserId
+          user: serviceUserId
         }).then((response) => {
           return Promise.all([
             test.expect(response.statusCode).to.equal(200),
@@ -216,7 +216,7 @@ test.describe('security tokens api', function () {
       test.it('should return a forbidden error if requested type if different to "apiKey"', () => {
         return getAuthTokens({
           type: 'jwt',
-          user: moduleUserId
+          user: serviceUserId
         }).then((response) => {
           return Promise.all([
             test.expect(response.body.message).to.contain('Not authorized'),